+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 * [api]: Added method to define custom `q.isa()` methods and implemented a method to type SARS-CoV-2 genomes using [constellations](https://github.com/cov-lineages/constellations) of mutations (the SARS-CoV-2 typer should be considered experimental at this stage and is not guaranteed to work in all cases) (0.3.0.dev9).
 * [api]: Loading both VCFs and Kmer sketches listed in the same file (0.3.0.dev10).
 * [analysis]: Changed default option to `--use-conda` for analysis.
-* [cli]: Added general query command for multiple types of queries (`gdi query hasa:feature` or `gdi query isa:sample`).
+* [cli]: Added general query command for multiple types of queries (`gdi query hasa:feature` or `gdi query isa:sample`) as well as summarizing features (`--features-summary`).
 
 # 0.2.0